+++ test.js
@@ -1,0 +1,19 @@
+'use strict'
+
+/**
+ * Dependencies
+ */
+
+const assert = require('assert')
+const schema_migrations = require('./index')
+
+/**
+ * Assertions
+ */
+
+describe("struct", () => {
+  it("should have tests", () => {
+    assert.ok(true)
+  })
+
+})